requirements.txt
docx
pypdf
streamlit
google-generativeai